Fix incompatibilities with MariaDB > 10.1
[finishes #176421895]

+++ db/migrations/20171013223336_remove_deprecated_v1_service_fields.rb
@@ -1,7 +1,9 @@
 Sequel.migration do
   up do
     alter_table :services do
-      drop_column :provider
+      if @db.class.to_s.include? 'Mysql'
+        drop_column :provider
+      end
       drop_column :url
       drop_column :version
     end
